Improved texture compression tool and started a resource packer that analyzes model files and compresses it’s textures as needed.

+++ Tools/TextureCompression/convert.py
@@ -34,10 +34,28 @@ def prepare():
 
 
 def main():
-    if len(sys.argv) < 3:
-        print 'python convert.py input.png'
+    if len(sys.argv) < 2:
+        print 'python convert.py input.png [output (with option extension for a specific format)]'
+        return
+
+    supportedFileExtensions = ['.png', '.dds', '.astc']
+    requestedFileExtensions = supportedFileExtensions
 
     inputFileName, inputFileExtension = os.path.splitext(sys.argv[1])
+    if inputFileExtension != '.png':
+        print 'Currently only png files are supported as input'
+        return
+
+    if len(sys.argv) >= 3:
+        outputFileName, outputFileExtension = os.path.splitext(sys.argv[2])
+        if outputFileExtension:
+            if outputFileExtension in supportedFileExtensions:
+                requestedFileExtensions = [outputFileExtension]
+            else:
+                print 'Specified output files extension is not supported: ' + outputFileExtension + ' (needs to be one of ' + str(supportedFileExtensions) + ')'
+                return
+    else:
+        outputFileName = inputFileName
 
     astcencPath = os.path.dirname(sys.argv[0])
     nvTextureToolsPath = os.path.dirname(sys.argv[0])
@@ -51,39 +69,43 @@ def main():
         print 'Script needs to be updated with ShaderConductor path for platform: ' + platform.system()
         return
 
-    #Convert to BCn format in DDS container for desktop platforms
-    subprocess.call([nvTextureToolsPath, '-dds10', '-srgb', '-bc1', inputFileName + inputFileExtension, inputFileName + '.dds'])
+    if '.png' in requestedFileExtensions:
+        if inputFileName + inputFileExtension != outputFileName + '.png':
+            shutil.copy2(inputFileName + inputFileExtension, outputFileName + '.png')
 
-    #Convert to ASTC format for mobile platforms
-    width = subprocess.check_output(['identify', '-format', '%[fx:w]', sys.argv[1]])
-    height = subprocess.check_output(['identify', '-format', '%[fx:h]', sys.argv[1]])
-    bitdepth = subprocess.check_output(['identify', '-format', '%[fx:z]', sys.argv[1]])
+    if '.dds' in requestedFileExtensions:
+        #Convert to BCn format in DDS container for desktop platforms
+        subprocess.call([nvTextureToolsPath, '-dds10', '-srgb', '-bc1', inputFileName + inputFileExtension, outputFileName + '.dds'])
 
-    width = int(width)
-    height = int(height)
-    bitdepth = int(bitdepth)
-    numLevels = int(1 + math.floor(math.log(max(width, height), 2)))
+    if '.astc' in requestedFileExtensions:
+        #Convert to ASTC format for mobile platforms
+        width = subprocess.check_output(['identify', '-format', '%[fx:w]', inputFileName + inputFileExtension])
+        height = subprocess.check_output(['identify', '-format', '%[fx:h]', inputFileName + inputFileExtension])
+        bitdepth = subprocess.check_output(['identify', '-format', '%[fx:z]', inputFileName + inputFileExtension])
 
-    print 'Number of mipmap levels: ' + str(numLevels)
+        width = int(width)
+        height = int(height)
+        bitdepth = int(bitdepth)
+        numLevels = int(1 + math.floor(math.log(max(width, height), 2)))
 
-    shutil.copy2(sys.argv[1], inputFileName + '.0' + inputFileExtension)
-    for i in range(1, numLevels):
-        subprocess.call(['convert', inputFileName + '.' + str(i-1) + inputFileExtension, '-scale', '50%', '-define', 'png:bit-depth='+str(bitdepth), inputFileName + '.' + str(i) + inputFileExtension])
+        print 'Number of mipmap levels: ' + str(numLevels)
 
-    for i in range(0, numLevels):
-        subprocess.call([astcencPath, '-c', inputFileName + '.' + str(i) + inputFileExtension, inputFileName + '.' + str(i) + '.astc', '6x6', '-medium'])
-        os.remove(inputFileName + '.' + str(i) + inputFileExtension)
+        shutil.copy2(inputFileName + inputFileExtension, outputFileName + '.0' + inputFileExtension)
+        for i in range(1, numLevels):
+            subprocess.call(['convert', outputFileName + '.' + str(i-1) + inputFileExtension, '-scale', '50%', '-define', 'png:bit-depth='+str(bitdepth), outputFileName + '.' + str(i) + inputFileExtension])
 
-
-    #subprocess.call(['sh', os.path.basename(compressonatorPath), '-fd', 'ASTC', sys.argv[1], sys.argv[2]], cwd=os.path.abspath(os.path.dirname(compressonatorPath)))
-
-    with open(inputFileName + '.astc', 'wb') as outputFile:
         for i in range(0, numLevels):
-            tempSourceFile = inputFileName + '.' + str(i) + '.astc'
-            if os.path.isfile(tempSourceFile):
-                with open(tempSourceFile, 'rb') as source:
-                    outputFile.write(source.read())
-                os.remove(tempSourceFile)
+            #subprocess.call(['sh', os.path.basename(compressonatorPath), '-fd', 'ASTC', sys.argv[1], sys.argv[2]], cwd=os.path.abspath(os.path.dirname(compressonatorPath)))
+            subprocess.call([astcencPath, '-c', outputFileName + '.' + str(i) + inputFileExtension, outputFileName + '.' + str(i) + '.astc', '6x6', '-medium'])
+            os.remove(outputFileName + '.' + str(i) + inputFileExtension)
+            
+        with open(outputFileName + '.astc', 'wb') as outputFile:
+            for i in range(0, numLevels):
+                tempSourceFile = outputFileName + '.' + str(i) + '.astc'
+                if os.path.isfile(tempSourceFile):
+                    with open(tempSourceFile, 'rb') as source:
+                        outputFile.write(source.read())
+                    os.remove(tempSourceFile)
 
 if __name__ == '__main__':
     prepare()
